Fix yolov8 pytorch notebook
